# src/main/resources/static/openapi.yaml
openapi: 3.1.0
info:
  title: DCAT-AP-SE Processor API
  version: 0.1.0
  x-api-released: '2022-10-01'
  x-api-status: beta
  summary: Converts API specifications into DCAT-AP-SE RDF/XML.
  description: |
    A tool for automating the production of metadata according to the
    [DCAT-AP-SE](https://docs.dataportal.se/dcat/en/) metadata specification,
    used to publish information about datasets and APIs on
    [Sveriges dataportal](https://www.dataportal.se/)
    The processor takes an API definition as input and generates an RDF file
    ready to be harvested by Sveriges dataportal

    ### Supported input
    - **OpenAPI** in JSON format, OAS2 and OAS3
    - **YAML/RAML** — 0.8 and 1.0
    - **Standalone metadata files** JSON/RAML/YAML for APIs without a formal definition

    ### Output
    RDF/XML conforming to the DCAT-AP-SE specification

    ### Usage
    Designed to be integrated into a CI/CD pipeline or run standalone, so that
    metadata stays in sync with the API definition it describes.
  license:
    name: EUPL-1.2
    url: https://joinup.ec.europa.eu/collection/eupl/eupl-text-eupl-12
externalDocs:
  description: DCAT-AP-SE-processor repository
  url: https://github.com/diggsweden/DCAT-AP-SE-Processor
paths:
  /dcat-generation/spec:
    post:
      tags:
        - generation
      operationId: restEndpointProduceRdfFromSpecRequest
      summary: Generate DCAT-AP-SE metadata from supplied specifications.
      description: |
        Accepts one or more API specifications in the request body and returns the
        resulting DCAT-AP-SE metadata as RDF/XML.
        The file extension in `name` selects the parser, so it must be present and
        match the format of `content`. Supported: `JSON`, `YAML` and `RAML`.
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpecRequest"
        required: true
      responses:
        "200":
          description: RDF/XML was generated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DcatResult"
        "400":
          description: The request body is invalid. The DcatResult.errorReport lists each problem found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DcatResult"
        "422":
          description: The specifications were processed but produced no valid metadata. The Error report is in DcatResult.errorReport.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DcatResult"
        "500":
          description: RDF generation failed. The Error report is in DcatResult.errorReport.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DcatResult"
  /dcat-generation/files/:
    get:
      tags:
        - generation
      operationId: generateRdfFromDirectory
      summary: Generate DCAT-AP-SE metadata from a directory.
      description: |
        Reads every API specification in the given directory and returns the
        resulting DCAT-AP-SE metadata as RDF/XML.

        The directory is resolved on the server running the processor, not on the client.
      parameters:
        - name: dir
          in: query
          required: false
          description: Server-side directory containing the API specifications to process.
          schema:
            type: string
            default: /apidef
      responses:
        "200":
          description: RDF/XML was generated, or a processing error message was returned.
          content:
            text/plain;charset=UTF-8:
              schema:
                type: string
  /dcat-generation/files/report:
    get:
      tags:
        - generation
      operationId: restEndpointProduceRdfAndQualityReport
      summary: Generate DCAT-AP-SE metadata and a quality report from a directory.
      description: |
        Reads every API specification in the given directory and returns the
        resulting DCAT-AP-SE metadata as RDF/XML, together with an error report
        and a quality report.

        The directory is resolved on the server running the processor, not on the client.
      parameters:
        - name: dir
          in: query
          required: false
          description: Server-side directory containing the API specifications to process.
          schema:
            type: string
            default: /apidef
      responses:
        "200":
          description: Processing finished. Check `errorReport` and `qualityReport` for problems.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DcatResult"
  /api-info:
    get:
      tags:
        - metadata
      operationId: getApiInfo
      summary: Get API information.
      description: Returns name, semantic version, release date, documentation URL and lifecycle status.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiInfo"
              examples:
                default:
                  value:
                    apiName: DCAT-AP-SE Processor API
                    apiVersion: 0.1.0
                    apiReleased: '2026-08-06'
                    apiDocumentation: https://example.org/swagger-ui
                    apiStatus: beta
components:
  schemas:
    ApiInfo:
      type: object
      properties:
        apiName:
          type: string
        apiVersion:
          type: string
        apiReleased:
          type: string
        apiDocumentation:
          type: string
        apiStatus:
          type: string
    ApiSource:
      type: object
      description: |
        A single API specification, supplied as a file name and its raw content.
      required: [name, content]
      properties:
        name:
          type: string
          description: |
            File name including extension. The extension selects the parser and must
            match the format of `content`. Supported: `.json`, `.yaml`, `.yml`, `.raml`.
        content:
          type: string
          description: |
            The specification as raw text.
    SpecRequest:
      type: object
      description: |
        A batch of API specifications to convert in a single request. All sources
        in one request are processed together and produce a single RDF/XML document,
        so specifications that reference each other should be sent together rather
        than in separate calls.
      required: [sources]
      properties:
        sources:
          type: array
          description: |
            The specifications (ApiSource) to convert. At least one is required. Each entry is
            validated before any processing starts, and a single invalid entry fails
            the whole request with 400 — nothing is partially converted.
          items:
            $ref: "#/components/schemas/ApiSource"
    DcatResult:
      type: object
      description: |
        The outcome of a generation run.
      properties:
        rdf:
          description: |
            The generated DCAT-AP-SE metadata as RDF/XML.
          type: string
        errorReport:
          description: |
            Errors found while validating the request or processing the specifications.
          type: string
        qualityReport:
          description: |
            Quality check of the generated metadata, grouped by DCAT-AP-SE main class.
          $ref: "#/components/schemas/QualityReport"
    MainClassQuality:
      type: object
      description: |
        Quality result for a single resource of a DCAT-AP-SE main class.
      properties:
        mainClass:
          description: |
            The DCAT-AP-SE main class, for example Catalog, Dataset or DataService.
          type: string
        uri:
          description: |
            URI of the resource that was checked.
          type: string
        issues:
          description: |
            Quality issues found for the resource. Empty if none were found.
          type: array
          items:
            $ref: "#/components/schemas/QualityIssue"
    QualityIssue:
      type: object
      description: |
        A single quality issue on a resource.
      properties:
        property:
          description: |
            The property the issue concerns, for example dcterms:publisher.
          type: string
        message:
          description: |
            Description of the issue.
          type: string
    QualityReport:
      type: object
      description: |
        Quality check of the generated metadata, grouped by DCAT-AP-SE main class.
      properties:
        completed:
          description: |
            Whether the quality check ran to completion. If false, the report is incomplete.
          type: boolean
        mainClasses:
          description: |
            Quality results per checked resource.
          type: array
          items:
            $ref: "#/components/schemas/MainClassQuality"